fix: inaccurate pearl data
The gravity, drag and the height of pearls is stored as float types, using double, to store them may casue some minor errors.

+++ src/main/java/com/github/zly2006/reden/pearl/MyEndPearlEntity.kt
@@ -6,13 +6,13 @@ class MyEndPearlEntity {
         const val width = 0.25
     }
 
-    val eyeY get() = height * 0.85
+    val eyeY get() = height * 0.85F
     val motion = Mutable3d()
     val pos = Mutable3d()
 
     fun tick() {
         pos.add(motion)
-        motion.multiply(0.99)
-        motion.add(0.0, -0.03, 0.0)
+        motion.multiply(0.99F.toDouble())
+        motion.add(0.0, -0.03F.toDouble(), 0.0)
     }
 }
